Added Improved Support for SQL SProcs
Added overrides to CreateCommand, Execute, Query, and Scalar that add
support for SQL Stored Procedures, Named Parameters, and SqlParameter.
Named parameters are prefered for use with stored procedures as they
don't break if the order changes.  Add support to ensure commandtype is
set correctly and a returnValue is added to the params.