Fixed Null References
Changed double if statements in ValidatesPresenceOf into an if/else
statement, because if the value is null it would through a
NullReferenceException when it called ToString on it.

In ValidateIsCurrency added a return after value == null check is true
as we don't need to process the rest of the method and if you did and
value was null he ToString call would throw a NullReferenceException.

In oracle changed a if _sequence is  not empty sring check to use be
!string.IsNullOrEmpty otherwise if a null _sequence it would still enter
the block.