Added Missing using Statement
Using statement is missing, this should be disposing the reader when
done. It is in sycn with the other Query method now as well.

Basically if you don't call Close on the Reader nothing will including
the finalizer. According to the MSDN doc's if you own the reader you
must explicitly call Close when you are through with it. The close
method has some clean up it does as described in the
SqlDataReader.Close.aspx) docs.